requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
redis==5.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
structlog==23.2.0
python-dotenv==1.0.0
langchain==0.0.350
langchain-google-genai==0.0.6
# Use newer Splunk SDK version that supports Python 3.13
splunk-sdk==2.0.2